Calculate anomaly score for the attack detector
