+++ malicious-smart-contract-ml-v3-py/.gitignore
@@ -4,3 +4,4 @@ forta.config.json
 __pycache__
 .pytest_cache
 malicious_non_token_model_11_05_22.joblib
+src/luabase_constants.py
