+++ malicious-smart-contract-ml-v3-py/src/agent.py
@@ -1,23 +1,21 @@
 import forta_agent
-import rlp
 from forta_agent import get_json_rpc_url
-from hexbytes import HexBytes
 from joblib import load
 from evmdasm import EvmBytecode
 from web3 import Web3
 
 from src.constants import (
     BYTE_CODE_LENGTH_THRESHOLD,
-    CONTRACT_SLOT_ANALYSIS_DEPTH,
     MODEL_THRESHOLD,
-    ERC721_SIGHASHES,
-    ERC20_SIGHASHES,
-    ERC1155_SIGHASHES,
-    ERC777_SIGHASHES,
-    PROXY_SIGHASHES,
 )
 from src.findings import MaliciousTokenContractFindings
 from src.logger import logger
+from src.utils import (
+    calc_contract_address,
+    get_anomaly_score,
+    get_features,
+    get_storage_addresses,
+)
 
 
 web3 = Web3(Web3.HTTPProvider(get_json_rpc_url()))
@@ -32,96 +30,6 @@ def initialize():
     logger.info("Start loading model")
     ML_MODEL = load("malicious_non_token_model_11_05_22.joblib")
     logger.info("Complete loading model")
-
-
-def is_contract(w3, address) -> bool:
-    """
-    this function determines whether address is a contract
-    :return: is_contract: bool
-    """
-    if address is None:
-        return True
-    code = w3.eth.get_code(Web3.toChecksumAddress(address))
-    return code != HexBytes("0x")
-
-
-def get_storage_addresses(w3, address) -> set:
-    """
-    this function returns the addresses that are references in the storage of a contract (first CONTRACT_SLOT_ANALYSIS_DEPTH slots)
-    :return: address_list: list (only returning contract addresses)
-    """
-    if address is None:
-        return set()
-
-    address_set = set()
-    for i in range(CONTRACT_SLOT_ANALYSIS_DEPTH):
-        mem = w3.eth.get_storage_at(Web3.toChecksumAddress(address), i)
-        if mem != HexBytes(
-            "0x0000000000000000000000000000000000000000000000000000000000000000"
-        ):
-            # looking at both areas of the storage slot as - depending on packing - the address could be at the beginning or the end.
-            addr_on_left = mem[0:20].hex()
-            addr_on_right = mem[12:].hex()
-            if is_contract(w3, addr_on_left):
-                address_set.add(Web3.toChecksumAddress(addr_on_left))
-            if is_contract(w3, addr_on_right):
-                address_set.add(Web3.toChecksumAddress(addr_on_right))
-
-    return address_set
-
-
-def get_contract_type(opcodes: str, function_sighashes: set) -> str:
-    """
-    this function determines contract type based on available sighashes.
-    :return: contract_type: str
-    """
-    if function_sighashes.intersection(ERC777_SIGHASHES):
-        return "erc777"
-    elif function_sighashes.intersection(ERC1155_SIGHASHES):
-        return "erc1155"
-    elif function_sighashes.intersection(ERC721_SIGHASHES):
-        return "erc721"
-    elif function_sighashes.intersection(ERC20_SIGHASHES):
-        return "erc20"
-    elif (
-        function_sighashes.intersection(PROXY_SIGHASHES) or len(function_sighashes) == 0
-    ):  # minimal proxy contract
-        return "proxy"
-    else:
-        return "non-token-or-proxy"
-
-
-def get_features(w3, opcodes) -> list:
-    """
-    this function returns the opcodes + function hashes contained in the contract
-    :return: features: list
-    """
-    features = []
-    function_sighashes = set()
-    opcode_addresses = set()
-
-    for i, opcode in enumerate(opcodes):
-        opcode_name = opcode.name
-        # treat unique unknown and invalid opcodes as UNKNOWN OR INVALID
-        if opcode_name.startswith("UNKNOWN") or opcode_name.startswith("INVALID"):
-            opcode_name = opcode.name.split("_")[0]
-        features.append(opcode_name)
-        if len(opcode.operand) == 40 and is_contract(w3, opcode.operand):
-            opcode_addresses.add(Web3.toChecksumAddress(f"0x{opcode.operand}"))
-
-        if i < (len(opcodes) - 3):
-            if (
-                opcodes[i].name == "PUSH4"
-                and opcodes[i + 1].name == "EQ"
-                and opcodes[i + 2].name == "PUSH2"
-                and opcodes[i + 3].name == "JUMPI"
-            ):  # add function sighashes
-                features.append(opcode.operand)
-                function_sighashes.add(f"0x{opcode.operand}")
-    features = " ".join(features)
-    contract_type = get_contract_type(features, function_sighashes)
-
-    return features, opcode_addresses, contract_type, function_sighashes
 
 
 def exec_model(w3, opcodes: str) -> tuple:
@@ -214,6 +122,7 @@ def detect_malicious_contract(w3, from_, created_contract_address, code) -> list
                 f"{created_contract_address}: type={contract_type}, score={model_score}, func_hashes={function_sighashes}"
             )
             if model_score is not None and model_score >= MODEL_THRESHOLD:
+                anomaly_score = get_anomaly_score(w3.eth.chain_id)
                 findings.append(
                     MaliciousTokenContractFindings.malicious_contract_creation(
                         from_,
@@ -222,20 +131,11 @@ def detect_malicious_contract(w3, from_, created_contract_address, code) -> list
                         function_sighashes,
                         model_score,
                         MODEL_THRESHOLD,
+                        anomaly_score,
                     )
                 )
 
     return findings
-
-
-def calc_contract_address(w3, address, nonce) -> str:
-    """
-    this function calculates the contract address from sender/nonce
-    :return: contract address: str
-    """
-
-    address_bytes = bytes.fromhex(address[2:].lower())
-    return Web3.toChecksumAddress(Web3.keccak(rlp.encode([address_bytes, nonce]))[-20:])
 
 
 def provide_handle_transaction(w3):
