+++ malicious-smart-contract-ml-v3-py/src/findings.py
@@ -10,6 +10,7 @@ class MaliciousTokenContractFindings:
         function_sighashes: set,
         model_score: float,
         model_threshold: float,
+        anomaly_score: float,
     ) -> Finding:
         metadata = {
             "address_contained_in_created_contract_" + str(i): address
@@ -22,6 +23,7 @@ class MaliciousTokenContractFindings:
         ] = f"https://oko.palkeo.com/{contract_address}/"
         metadata["model_score"] = str(model_score)
         metadata["model_threshold"] = str(model_threshold)
+        metadata["anomaly_score"] = anomaly_score
 
         return Finding(
             {
